+++ Kahla.Server/Controllers/GroupsController.cs
@@ -23,6 +23,7 @@ namespace Kahla.Server.Controllers
         private readonly UserManager<KahlaUser> _userManager;
         private readonly KahlaDbContext _dbContext;
         private readonly KahlaPushService _pusher;
+        private static readonly object _obj = new object();
 
         public GroupsController(
             UserManager<KahlaUser> userManager,
@@ -91,27 +92,107 @@ namespace Kahla.Server.Controllers
             {
                 return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: {groupName}!");
             }
-            var joined = group.Users.Any(t => t.UserId == user.Id);
-            if (joined)
-            {
-                return this.Protocol(ErrorType.HasDoneAlready, $"You have already joined the group: {groupName}!");
-            }
             if (group.HasPassword && group.JoinPassword != joinPassword?.Trim())
             {
                 return this.Protocol(ErrorType.WrongKey, "The group requires password and your password was not correct!");
             }
-            // All checked and able to join him.
-            // Warning: Currently we do not have invitation system for invitation control is too complicated.
-            var newRelationship = new UserGroupRelation
+            lock (_obj)
             {
-                UserId = user.Id,
-                GroupId = group.Id
-            };
-            _dbContext.UserGroupRelations.Add(newRelationship);
-            await _dbContext.SaveChangesAsync();
+                var joined = group.Users.Any(t => t.UserId == user.Id);
+                if (joined)
+                {
+                    return this.Protocol(ErrorType.HasDoneAlready, $"You have already joined the group: {groupName}!");
+                }
+                // All checked and able to join him.
+                // Warning: Currently we do not have invitation system for invitation control is too complicated.
+                var newRelationship = new UserGroupRelation
+                {
+                    UserId = user.Id,
+                    GroupId = group.Id
+                };
+                _dbContext.UserGroupRelations.Add(newRelationship);
+                _dbContext.SaveChanges();
+            }
             await group.ForEachUserAsync((eachUser, relation) => _pusher.NewMemberEvent(eachUser, user, group.Id), _userManager);
             return this.Protocol(ErrorType.Success, $"You have successfully joint the group: {groupName}!");
         }
+
+        [HttpPost]
+        public async Task<IActionResult> TransferGroupOwner([Required]string groupName, [Required]string targetUserId)
+        {
+            var user = await GetKahlaUser();
+            var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == groupName);
+            if (group == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: '{groupName}'!");
+            }
+            if (group.OwnerId != user.Id)
+            {
+                return this.Protocol(ErrorType.Unauthorized, $"You are not the owner of this group: '{groupName}' and you can't transfer it!");
+            }
+            // current user is the owner of the group.
+            var targetRelationship = await _dbContext.GetRelationFromGroup(targetUserId, group.Id);
+            if (targetRelationship == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"We can not find the target user with id: '{targetUserId}' in the group with name: '{groupName}'!");
+            }
+            if (group.OwnerId == targetUserId)
+            {
+                return this.Protocol(ErrorType.RequireAttention, $"Caution! You are already the owner of the group '{groupName}'.");
+            }
+            group.OwnerId = targetUserId;
+            await _dbContext.SaveChangesAsync();
+            return this.Protocol(ErrorType.Success, $"Successfully transfered your group '{groupName}' ownership!");
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> KickMember([Required]string groupName, [Required]string targetUserId)
+        {
+            var user = await GetKahlaUser();
+
+            var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == groupName);
+            if (group == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: '{groupName}'!");
+            }
+            if (group.OwnerId != user.Id)
+            {
+                return this.Protocol(ErrorType.Unauthorized, $"You are not the owner of this group: '{groupName}' and you can't kick out members.!");
+            }
+            var targetuser = await _dbContext
+                .UserGroupRelations
+                .SingleOrDefaultAsync(t => t.GroupId == group.Id && t.UserId == targetUserId);
+            if (targetuser == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"We can not find the target user with id: '{targetUserId}' in the group with name: '{groupName}'!");
+            }
+            _dbContext.UserGroupRelations.Remove(targetuser);
+            await group.ForEachUserAsync((eachUser, relation) => _pusher.SomeoneLeftEvent(eachUser, targetuser.User, group.Id), _userManager);
+            await _dbContext.SaveChangesAsync();
+            return this.Protocol(ErrorType.Success, $"Successfully kicked the member from group '{groupName}'.");
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> DissolveGroup([Required]string groupName)
+        {
+            var user = await GetKahlaUser();
+
+            var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == groupName);
+            if (group == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: '{groupName}'!");
+            }
+            if (group.OwnerId != user.Id)
+            {
+                return this.Protocol(ErrorType.Unauthorized, $"You are not the owner of the group: '{groupName}' and you can't dissolve it.!");
+            }
+
+            await group.ForEachUserAsync((eachUser, relation) => _pusher.DissolveEvent(eachUser, group.Id), _userManager);
+            _dbContext.GroupConversations.Remove(group);
+            await _dbContext.SaveChangesAsync();
+            return this.Protocol(ErrorType.Success, $"Successfully dissolved the group '{groupName}'!");
+        }
+
 
         [HttpPost]
         public async Task<IActionResult> LeaveGroup([Required]string groupName)
@@ -120,12 +201,16 @@ namespace Kahla.Server.Controllers
             var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == groupName);
             if (group == null)
             {
-                return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: {groupName}!");
+                return this.Protocol(ErrorType.NotFound, $"We can not find a group with name: '{groupName}'!");
             }
             var joined = await _dbContext.GetRelationFromGroup(user.Id, group.Id);
             if (joined == null)
             {
-                return this.Protocol(ErrorType.HasDoneAlready, $"You did not joined the group: {groupName} at all!");
+                return this.Protocol(ErrorType.HasDoneAlready, $"You did not joined the group: '{groupName}' at all!");
+            }
+            if (group.OwnerId == user.Id)
+            {
+                return this.Protocol(ErrorType.NotEnoughResources, $"You are the owner of this group: '{groupName}' and you can't leave it!");
             }
             _dbContext.UserGroupRelations.Remove(joined);
             await _dbContext.SaveChangesAsync();
@@ -163,9 +248,60 @@ namespace Kahla.Server.Controllers
             return this.Protocol(ErrorType.Success, $"Successfully {(setMuted ? "muted" : "unmuted")} the group '{groupName}'!");
         }
 
-        private Task<KahlaUser> GetKahlaUser()
+        [HttpPost]
+        public async Task<IActionResult> UpdateGroupInfo(UpdateGroupAddressModel model)
         {
-            return _userManager.GetUserAsync(User);
+            var user = await GetKahlaUser();
+            var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == model.GroupName);
+            if (group == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"Can not find a group with name: {model.GroupName}!");
+            }
+            if (group.OwnerId != user.Id)
+            {
+                return this.Protocol(ErrorType.Unauthorized, "You can't change settings for this group for you are not the owner of it.");
+            }
+
+            if (!string.IsNullOrEmpty(model.NewName))
+            {
+                model.NewName = model.NewName.Trim().ToLower();
+                if (model.NewName != group.GroupName)
+                {
+                    if (_dbContext.GroupConversations.Any(t => t.GroupName == model.NewName))
+                    {
+                        return this.Protocol(ErrorType.NotEnoughResources, $"A group with name: '{model.NewName}' already exists!");
+                    }
+                    group.GroupName = model.NewName;
+                }
+            }
+
+            if (model.AvatarKey != null)
+            {
+                group.GroupImageKey = model.AvatarKey.Value;
+            }
+
+            await _dbContext.SaveChangesAsync();
+            return this.Protocol(ErrorType.Success, $"Successfully updated the name of the group '{model.GroupName}'.");
         }
+
+        [HttpPost]
+        public async Task<IActionResult> UpdateGroupPassword(UpdateGroupPasswordAddressModel model)
+        {
+            var user = await GetKahlaUser();
+            var group = await _dbContext.GroupConversations.SingleOrDefaultAsync(t => t.GroupName == model.GroupName);
+            if (group == null)
+            {
+                return this.Protocol(ErrorType.NotFound, $"Can not find a group with name: '{model.GroupName}'!");
+            }
+            if (group.OwnerId != user.Id)
+            {
+                return this.Protocol(ErrorType.Unauthorized, "You can't change settings for this group for you are not the owner of it.");
+            }
+            group.JoinPassword = model.NewJoinPassword;
+            await _dbContext.SaveChangesAsync();
+            return this.Protocol(ErrorType.Success, $"Successfully update the join password of the group '{model.GroupName}'.");
+        }
+
+        private Task<KahlaUser> GetKahlaUser() => _userManager.GetUserAsync(User);
     }
 }
